Fixing timestamp tag retrieval
Before this fix we passed the commit timestamp that was used to create a tag.
Instead, we need to use the creation timestamp of a tag, because it is
not the same as the timestamp of the last commit.

The other refactor here is that we can simplify processing of two latest tags.
keep-core client already has 2 versions and the versions will only be
added, not removed. So we don't need all the logic here checking for
number of versions. And rewards are always calculated by the end of the
month.