Changing the way we save build version in JSON file
It might happen that the same instance can run on different versions of
a client in a given rewards interval. Before we only saved the latest
version of a client, but now we save all the versions for a given
instance along with their last registered timestamps.
E.g.
"instances": {
	"3.145.42.190:9601": {
                (...)
                "version": {
                 "v2.0.0-m1": 1676073780, //last registered timestamp
		 "v2.0.0-m2": 1676548212  //last registered timestamp
                }
        }
}