Replace the tags fetching to create temp remotes
The tBTCv2 node client releases are collected fetching the git tags
created for each release in keep-core repo.

Up to now, the strategy to fetch these tags located on a different repo
is to set the URL of the `origin` remote to keep-core's one and run
`git fetch`. This strategy may lead to several problems:

- If the script fails before the statement that returns the original URL
  to `origin` remote, this remote will keep the new keep-core repo URL,
  which can cause future troubles.
- If you have a different configuration of remotes without a remote
  called `origin`, the tBTCv2 rewards calculation script fails.

The proposed changes modify the strategy: a new remote is created with
keep-core URL and tags are fetched from there. After calculating the
rewards, this new remote is removed.