New Merkle Distribution for Mar 1st
As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

The rewards earned in this distribution are are determined for the period
between Feb 1st and Mar 1st 2023 and include:

- PRE rewards for those who have a PRE operator confirmed and were
  staking during this period.
- tBTCv2 rewards for those who comply with the tBTCv2 requirements
  during this period.

The reward allocation for this distribution is:

- PRE rewards: 37.5%
- tBTCv2 rewards: 62.5%

Note that since this is the first distribution in which a version for
tBTCv2 operator client has been released, the upgrade window has been
increased to let the operators more time to do this first update.
Operators was allowed to perform the upgrade until Feb 28 00:00:00 UTC.