Lowering uptime req to 50%
There is an issue with a keep client that might result in errors that
affect uptime requirement in rewards distribution. KEEP team decided to
make an exception for the month of Aug and Sep and lower the uptime req
to 50%. It should be enough that every single node meet the uptime req
for the month of Aug.