Validating all the keep-client versions over the rewards interval
This is a follow up fix of PR #78. In both cases a script should
validate all the keep-client versions in a given rewards interval. If it
happens that a client run an older version that is allowed it should
not be eligible for rewards.

In this scenario when a cutoff day for the upgrade enters the next
rewards distribution a staker can run either of two latest client
versions at any given moment.