Adjusting uptime req for the October interval
In a new version of Keep client -m5 there was a fix that mitigated an
issue that might've affected the uptime of the client. The cutoff day
to update a client is set to Oct 11th which means that a staker can
still experience the issue with the uptime. To offset that, we decided
to keep the bar lower of the uptime and calculate it the following way:
11 - cutoff day
31 - days in Oct
4% - default allowed downtime
11/31 - 4% gives around 61%. We can lower it to 60%. This way it gives
slack for the users that can still experience uptime issues.

For Nov interval the uptime req should return to its original value of
96%