Adding an uptime check for the entire interval
There was a check that validated an uptime for a given instance but only
when an instance was running. A node can spin up and start running only
e.g. for 10 days and report 100% uptime, but it was just around 33% of the
entire interval. In order to be eligilbe for rewards, a node should run at least a
"requiredUptime" of an interval.
In this commit we check not only an uptime of a node while it was
running, but also a percentage of an interval it was running in. Both values need to
be >= "requiredUptime" to get the rewards.