Adding support for a third eligible release in a given interval
So far for simplicity purposes we supported only 2 keep client release
tags in a given interval. In a rare cases when we need to release a hot
fix it might happen that in a given interval 3 keep client versions will
be eligible, just like for October 23 inteval. This commit adds a
support for a third release. However, the default number of tags should
be set to 2, only in rare cases the number should be bumped up to 3, ie
when 3 tags are eligible to run in a given interval.