Include `m6` announcement drift within the allowed upgrade delay
The `m6` release was announced 20 days after the tag was published on GitHub.
That means we need to extend the default allowed upgrade delay (3 weeks)
to take that additional delay into account. The new allowed upgrade
delay is 41 days. That means the cutoff day for `m6` is November 15.
The modified allowed upgrade delay should remain in force for October and
November rewards intervals. We should bring it back to the default value
of 3 weeks for December interval.