New Merkle Distribution for Nov 1st 23
As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

The rewards earned in this distribution are determined for the period
between Oct 1st and Nov 1st 2023 and include:

- PRE rewards for those who have a PRE operator confirmed and were
  staking during this period.
- tBTCv2 rewards for those who comply with the tBTCv2 requirements
  during this period.

The reward allocation for this distribution is:

- PRE rewards: 25%
- tBTCv2 rewards: 75%

Note: for this specific distribution, the tBTCv2 client update time
window has been increased since the announcement of the new version of
the client v2.0.0-m6 was officially made aprox 23 days after the
release. So, for this time, the rewards calculation script considers
that the announcement date is the starting date, and not the actual
release date. More info in PR #91.

Note: for this specific distribution, the uptime requirement for tBTCv2
client has been set from 60%. The reason for this is an issue in Keep
client. See PR #76.