New Merkle Distribution for Nov 22nd 23
As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

The rewards earned in this distribution are determined for the period
between Nov 1st 00:00:00 and Nov 22nd 04:06:11 2023 and include:

- PRE rewards for those who have a PRE operator confirmed and were
  staking during this period.
- tBTCv2 rewards for those who comply with the tBTCv2 requirements
  during this period.

The reward allocation for this distribution is:

- PRE rewards: 25%
- tBTCv2 rewards: 75%

Note: tBTCv2 uptime requirement has been returned to default value 96%.
More info in PR #104.

Note: The period of this distribution has been shortened due to the
legacy stake dropped support. More info can be found in [1], [2]. The
transaction that disabled the legacy stakes was executed Nov 22nd 2023
04:06:11 UTC, so for technical reasons, the distribution end timestamp
is this exact time. The transaction can be found in [3].

[1] https://github.com/threshold-network/solidity-contracts/issues/141
[2] https://forum.threshold.network/t/transition-guide-for-legacy-stakers/719
[3] https://etherscan.io/tx/0x68ddee6b5651d5348a40555b0079b5066d05a63196e3832323afafae0095a656