New Merkle Distribution for Dec 1st 23
As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

The rewards earned in this distribution are determined for the period
between Nov 22nd 04:06:11 and Dec 1st 2023 and include:

- PRE rewards for those who have a PRE operator confirmed and were
  staking during this period.
- tBTCv2 rewards for those who comply with the tBTCv2 requirements
  during this period.

The reward allocation for this distribution is:

- PRE rewards: 25%
- tBTCv2 rewards: 75%

Note: The period of this distribution has been shortened due to the
legacy stake dropped support. More info can be found in [1], [2]. The
transaction that disabled the legacy stakes was executed Nov 22nd 2023
04:06:11 UTC, so for technical reasons, the previous distribution ended
at this exact time and hence, this distribution starts at this time.
The transaction can be found in [3].

Note: Since the deadline for the legacy tokens transition has not yet
ended (see [2]), and considering that stakes that complete the process
will earn the complete amount of rewards for their legacy tokens, but
those that won't complete the process in time won't, it is not possible
to know how much rewards will earn those stakes with legacy tokens. For
this reason, no rewards have been calculated for legacy stakes in the
period of this distribution. In distribution after transition deadline,
the remaining rewards will be distributed.

[1] https://github.com/threshold-network/solidity-contracts/issues/141
[2] https://forum.threshold.network/t/transition-guide-for-legacy-stakers/719
[3] https://etherscan.io/tx/0x68ddee6b5651d5348a40555b0079b5066d05a63196e3832323afafae0095a656