New Merkle Distribution for Dec 8th 23
As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

Note that the date of this distribution is the date of legacy Nu staking
transition deadline [1].

The rewards earned in this distribution include:

- tBTCv2 rewards for the stakes who comply with the tBTCv2 requirements
  for the period Dec 1st - Dec 8th 23. Note that stakes with legacy Keep
  staking will not be included in this distribution. Info: [1] [2].
- PRE rewards for those that have a PRE operator confirmed and were
  staking T for the period Dec 1st - Dec 8th 23.
- PRE rewards for those that had legacy Nu staking and completed the
  transition process. This is, stakes that took their unstaked nuInT
  and restaked it in the transition period. The rewards have been
  calculated as if no unstaking was made. Partial restaking have
  received pro-rata rewards. Since they didn't receive the rewards for
  the period Nov 22nd - Dec 1st [3], these rewards correspond to the
  period Nov 22nd - Dec 8th.
- PRE rewards for those that had legacy Nu staking and regular T staking
  and didn't complete the transition process. These stakes won't receive
  legacy staking rewards, but since they didn't receive staking rewards
  for their T staking during the period Nov 22nd - Dec 1st [3], these
  rewards will be distributed now.

The rewards allocation for this distribution is:

- PRE rewards: 25%
- tBTCv2 rewards: 75%

Note: The period of this distribution has been shortened due to the
legacy stake deactivation. More info can be found in [1], [4]. The last
distribution was made on Dec 1st and the legacy Nu transition process
deadline ended Dec 8th.

Note: Since the deadline for the legacy tokens transtion has not yet
ended for legacy Keep staking, and considering that stakes that
completed the process will earn the complete amount of rewards for their
legacy tokens, but those that won't complete the process in time won't,
it is not possible to know how much rewards have been calculated for
those stakes with legacy tokens. For this reason, no rewards have been
calculated for legacy stakes in this distribution period. In future
distributions after the transition deadline, the remaining rewards will
be distributed.

[1] https://forum.threshold.network/t/transition-guide-for-legacy-stakers/719
[2] https://github.com/threshold-network/merkle-distribution/issues/112
[3] https://github.com/threshold-network/merkle-distribution/pull/111
[4] https://github.com/threshold-network/solidity-contracts/issues/141