New Merkle Distribution for Jan 1st 24
As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

The rewards earned in this distribution are determined for the period
between Dec 8th 2023 and Jan 1st 2023 and include:

- tBTCv2 rewards for the stakes who comply with the tBTCv2 requirements
for the period Dec 8th 23 - Jan 1st 24. Note that stakes with legacy
Keep staking will not be included in this distribution. See note below.

- TACo/PRE rewards for those that have a PRE or TACo operator confirmed
and were staking T for the period Dec 8th 23 - Jan 1st 24. This is the
TACo operator transition period, so for the calculation of this
distribution rewards both PRE and TACo operators are valid [3]. Stakes
that updated their PRE operator to TACo are earning rewards since the
confirmation of the PRE operator, so no rewards are lost in the
migration.

Note: Since the deadline for the legacy tokens transtion has not yet
ended for legacy Keep staking, and considering that stakes that
completed the process will earn the complete amount of rewards for their
legacy tokens, but those that won't complete the process in time won't,
it is not possible to know how much rewards have been calculated for
those stakes with legacy tokens. For this reason, no rewards have been
calculated for legacy stakes in this distribution period. In future
distributions after the transition deadline, the remaining rewards will
be distributed. Info: [1] [2].

[1] https://forum.threshold.network/t/transition-guide-for-legacy-stakers/719
[2] https://github.com/threshold-network/merkle-distribution/issues/112
[3] https://github.com/threshold-network/merkle-distribution/issues/113