New Merkle Distribution for Feb 1st 24
As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

The rewards earned in this distribution are determined for the period
between Jan 1st 2024 and Feb 1st 2024 and include:

- One-off Commitment Bonus: this is a incentive for longer staking
  commitments at genesis. Stakes that did the lock-up extension before
  January 15th, 2024 (UTC 23:59:59) received an extra yield that varies
  according to the extra lock-up period choosen [1].

- TACo rewards: for those that have a TACo operator confirmed and were
  staking T and authorized TACo application for the period Jan 1st 24 -
  Feb 1st 24.

- tBTCv2 rewards: for those that complied with the tBTCv2 requirements
  for the period Jan 1st 24 - Feb 1st 24. Note that stakes with legacy
  Keep staking won't be included in this distribution. See note below.

The rewards allocation for this distribution is 25% for TACo rewards and
75% for tBTC rewards.

Note: No rewards for stakes with legacy Keep. Since the deadline for the
legacy tokens transtion has not yet ended for legacy Keep staking, and
considering that stakes which completed the process will earn the
complete amount of rewards for their legacy tokens at the end of the
process, but those that won't complete the process in time won't, it is
not possible to know how much rewards have been calculated for those
stakes with legacy tokens. For this reason, no rewards have been
calculated for legacy stakes in this distribution period. In future
distributions after the transition deadline, the remaining rewards will
be distributed. Info: [2] [3].

[1] https://docs.threshold.network/staking-and-running-a-node/taco-node-setup/taco-authorization-and-operator-registration/one-off-commitment-bonus
[2] https://forum.threshold.network/t/transition-guide-for-legacy-stakers/719
[3] #112