New Merkle distribution for Mar 1st 24
As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distribution.

The rewards earned in this distribution are determined for the period
between Feb 24th and Mar 1st 2024 and include:

- TACo rewards. Allocation: 25%.
- tBTC rewards. Allocation: 75%.

Note that the previoud release was a partial distribution for the period
Feb 1st - Feb 24th. So this distribution is shorter than usual.