Rename event emitted when claimMerkle is executed
The current name `Claimed` can lead to confusion since there are two
sources of rewards being claimed: the rewards that were calculated and
distributed using a Merkle distribution, and the rewards that are
distributed by the applications (TACo). The former will emit a
MerkleClaimed event, and the latter, a RewardsWithdrawn event.