New Merkle distribution for Feb 1st 25
As this is a cumulative Merkle distribution, the rewards here collected are the sum of the rewards distributed in previous Merkle distributions plus the rewards earned in the period between the last and the current distribution.

The rewards earned in this distribution are determined for the period between Jan 1st 25 and Feb 1st 25, and include:

TACo rewards. Allocation 25%.
tBTC rewards. Allocation 75%.